requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
speechrecognition>=3.10.0
pyaudio>=0.2.11
python-dotenv>=1.0.0
pydub>=0.25.1
edge-tts>=6.1.0